requirements.txt
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.4
click==8.0.1
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.8.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
libretranslatepy==2.1.1
lxml==4.6.3
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.1
pytesseract==0.3.8
requests==2.26.0
rfc3986==1.5.0
sniffio==1.2.0
translate==3.6.1
Unidecode==1.2.0
urllib3==1.26.6
